refactor: extract useProcessedWidgets composable, migrate NodeWidgets tests to VTL
Extract ~190 lines of processedWidgets computed logic from NodeWidgets.vue
into useProcessedWidgets composable. Unit test borderStyle/promotion logic
directly on the composable. Remove mountComponent()/getBorderStyles() VTU
helpers and the 2 wrapper.vm-based tests from component test. Component
test is now 100% VTL (14 tests remain).

Amp-Thread-ID: https://ampcode.com/threads/T-019d6e80-d70b-72e7-9e08-73c88c578823
